Improve sparc condcodexform_cc results by introducing Cfg::removeOrphanBBs
This method will walk the Cfg, and for every BasicBlock that has no InEdges and is not an entry BasicBlock, it removes them.
It does this untill no more orphan BasicBlocks can be removed.

Keep m_mapBB in sync with Cfg::m_listBB.

Remove initialization of BasicBlock Ord and RevOrd to help in tracking down uninitialized accesses to those 2 values

Fix a few undefined behaviours in ssl files ( those were doing (1<<32) and (1<<31) - 1 on signed integers)